Move cookieOptions to constans.js
