+++ src/constants.js
@@ -1,0 +1,7 @@
+const cookieOptions = {
+  httpOnly: true,
+  secure: true,
+  sameSite: "None",
+};
+
+export { cookieOptions };
