+++ src/controllers/deliveryPartner.controller.js
@@ -18,7 +18,7 @@ import {
 import jwt from "jsonwebtoken";
 import fs from "fs";
 import { uploadImageOnCloudinary } from "../utils/cloudinary.js";
-import { cookieOptions } from "../utils/cookieOptions.js";
+import { cookieOptions } from "../constants.js";
 
 const getDeliveryPartnerAccount = asyncHandler(async (req, res) => {
   if (!req.deliveryPartner || !req.deliveryPartner.id) {
