--- src/utils/cookieOptions.js
@@ -1,7 +1,0 @@
-const cookieOptions = {
-  httpOnly: true,
-  secure: true,
-  sameSite: "None",
-};
-
-export { cookieOptions };
